+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/integration/recipemod/rei/RefinedStorageREIClientPlugin.java
@@ -2,13 +2,13 @@ package com.refinedmods.refinedstorage2.platform.fabric.integration.recipemod.re
 
 import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.api.integration.recipemod.IngredientConverter;
+import com.refinedmods.refinedstorage2.platform.api.item.AbstractEnergyBlockItem;
 import com.refinedmods.refinedstorage2.platform.common.block.ColorableBlock;
 import com.refinedmods.refinedstorage2.platform.common.content.BlockColorMap;
 import com.refinedmods.refinedstorage2.platform.common.content.Blocks;
 import com.refinedmods.refinedstorage2.platform.common.content.ContentIds;
 import com.refinedmods.refinedstorage2.platform.common.content.Items;
 import com.refinedmods.refinedstorage2.platform.common.content.Tags;
-import com.refinedmods.refinedstorage2.platform.common.item.block.ControllerBlockItem;
 import com.refinedmods.refinedstorage2.platform.common.screen.AbstractBaseScreen;
 
 import java.util.function.Supplier;
@@ -73,7 +73,8 @@ public class RefinedStorageREIClientPlugin implements REIClientPlugin {
         registry.group(
             createIdentifier("fully_charged_controller"),
             createTranslation("block", "controller.rei_fully_charged"),
-            ControllerBlockItem.getAllAtCapacity().map(EntryStacks::of).collect(Collectors.toList())
+            AbstractEnergyBlockItem.createAllAtEnergyCapacity(Items.INSTANCE.getControllers())
+                .map(EntryStacks::of).collect(Collectors.toList())
         );
         groupItems(
             registry,